Fix single-quoting style in PDF manuals
The PDF versions of the GNU manuals used curved single quotes to
represent grave accent and apostrophe, which made it a pain to cut
and paste code examples from them.  Fix the PDF versions to use
grave accent and apostrophe for Lisp source code, keystrokes, etc.
This change does not affect the info files, nor does it affect
ordinary uses of curved single quotes in PDF.

* docstyle.texi: New file, which specifies treatment for grave accent
and apostrophe, as well as the document encoding.

* auth.texi, emacs-mime.texi, epa.texi, gnus-coding.texi:
* gnus-faq.texi, gnus.texi, message.texi, sasl.texi, sieve.texi:
Use it instead of '@documentencoding UTF-8', to lessen the need for
global changes like this in the future.